Refactor to allow mounting new embeds to be easier
