+++ embeds/image/ImageEmbed.js
@@ -1,0 +1,11 @@
+import { h } from 'preact';
+
+const ImageEmbed = ({ fields: { name }, updateFields }) => (
+  <input
+    type="text"
+    value={name}
+    onInput={e => updateFields({ name: e.target.value })}
+  />
+);
+
+export default ImageEmbed;
